Fix MSVC compiler warnings
reg. conversion from 'size_t' to 'int', possible loss of data (C4267)

+++ src/engraving/dom/glissando.cpp
@@ -387,8 +387,8 @@ Note* Glissando::guessFinalNote(Chord* chord, Note* startNote)
         }
         // normal case: try to return the note in the next chord that is in the
         // same position as the start note relative to the end chord
-        int startNoteIdx = mu::indexOf(chord->notes(), startNote);
-        int endNoteIdx = std::min(startNoteIdx, int(target->notes().size()) - 1);
+        size_t startNoteIdx = mu::indexOf(chord->notes(), startNote);
+        size_t endNoteIdx = std::min(startNoteIdx, target->notes().size() - 1);
         return target->notes().at(endNoteIdx);
     }
